Move Switch-related code that was left behind.
The call to this code was moved in
078ac53de2cf3788f1b8c0242bb9b143a84a1025, but the code itself was
not.
