+++ RSDKv5/main.cpp
@@ -1,31 +1,6 @@
 #include "RSDK/Core/RetroEngine.hpp"
 #include "main.hpp"
 
-#ifdef __SWITCH__
-#include <switch.h>
-#include <stdlib.h>
-#include <stdio.h>
-#include <string.h>
-#include <sys/socket.h>
-#include <sys/errno.h>
-#include <arpa/inet.h>
-#include <unistd.h>
-
-static int32 s_nxlinkSock = -1;
-
-static void initNxLink()
-{
-    if (R_FAILED(socketInitializeDefault()))
-        return;
-
-    s_nxlinkSock = nxlinkStdio();
-    if (s_nxlinkSock >= 0)
-        printf("printf output now goes to nxlink server\n");
-    else
-        socketExit();
-}
-#endif
-
 #if RETRO_STANDALONE
 
 #if RETRO_PLATFORM == RETRO_WIN && !RETRO_RENDERDEVICE_SDL2
